Move the ShipController code into ship/, split PlayerShipController into it's own file.
