+++ src/Player.h
@@ -7,9 +7,9 @@
 #include "HyperspaceCloud.h"
 #include "Ship.h"
 #include "ShipCockpit.h"
-#include "ShipController.h"
 #include "galaxy/StarSystem.h"
 #include "libs.h"
+#include "ship/PlayerShipController.h"
 #include <list>
 
 namespace Graphics {
