+++ src/ship/PlayerShipController.cpp
@@ -1,7 +1,7 @@
 // Copyright © 2008-2019 Pioneer Developers. See AUTHORS.txt for details
 // Licensed under the terms of the GPL v3. See licenses/GPL-3.txt
 
-#include "ShipController.h"
+#include "PlayerShipController.h"
 #include "Frame.h"
 #include "Game.h"
 #include "GameConfig.h"
@@ -14,13 +14,6 @@
 #include "Ship.h"
 #include "Space.h"
 #include "WorldView.h"
-
-void ShipController::StaticUpdate(float timeStep)
-{
-	OS::EnableFPE();
-	m_ship->AITimeStep(timeStep);
-	OS::DisableFPE();
-}
 
 PlayerShipController::PlayerShipController() :
 	ShipController(),
