+++ src/ship/PlayerShipController.h
@@ -1,61 +1,15 @@
 // Copyright © 2008-2019 Pioneer Developers. See AUTHORS.txt for details
 // Licensed under the terms of the GPL v3. See licenses/GPL-3.txt
 
-#ifndef _SHIPCONTROLLER_H
-#define _SHIPCONTROLLER_H
-/*
- * Ship movement controller class
- * Controls thrusters, autopilot according to player input or AI
- */
-#include "JsonFwd.h"
-#include "libs.h"
+#pragma once
+
+#include "Input.h"
+#include "ShipController.h"
 
 namespace KeyBindings {
 	struct ActionBinding;
 	struct AxisBinding;
 } // namespace KeyBindings
-
-class Body;
-class Ship;
-class Space;
-
-enum FlightControlState { // <enum scope='FlightControlState' name=ShipControllerFlightControlState public>
-	CONTROL_MANUAL,
-	CONTROL_FIXSPEED,
-	CONTROL_FIXHEADING_FORWARD,
-	CONTROL_FIXHEADING_BACKWARD,
-	CONTROL_FIXHEADING_NORMAL,
-	CONTROL_FIXHEADING_ANTINORMAL,
-	CONTROL_FIXHEADING_RADIALLY_INWARD,
-	CONTROL_FIXHEADING_RADIALLY_OUTWARD,
-	CONTROL_FIXHEADING_KILLROT,
-	CONTROL_AUTOPILOT,
-
-	CONTROL_STATE_COUNT // <enum skip>
-};
-
-// only AI
-class ShipController {
-public:
-	//needed for serialization
-	enum Type {
-		AI = 0,
-		PLAYER = 1
-	};
-	ShipController() {}
-	virtual ~ShipController() {}
-	virtual Type GetType() { return AI; }
-	virtual void SaveToJson(Json &jsonObj, Space *s) {}
-	virtual void LoadFromJson(const Json &jsonObj) {}
-	virtual void PostLoadFixup(Space *) {}
-	virtual void StaticUpdate(float timeStep);
-	virtual void SetFlightControlState(FlightControlState s) {}
-	virtual FlightControlState GetFlightControlState() const { return CONTROL_MANUAL; }
-	virtual double GetSetSpeed() const { return 0.0; }
-	virtual void ChangeSetSpeed(double delta) {}
-	virtual Body *GetSetSpeedTarget() const { return nullptr; }
-	Ship *m_ship;
-};
 
 // autopilot AI + input
 class PlayerShipController : public ShipController {
@@ -98,7 +52,7 @@ public:
 	sigc::signal<void> onRotationDampingChanged;
 
 private:
-	static struct InputBinding {
+	static struct InputBinding : public Input::InputFrame {
 		// We create a local alias for ease of typing these bindings.
 		typedef KeyBindings::AxisBinding AxisBinding;
 		typedef KeyBindings::ActionBinding ActionBinding;
@@ -156,5 +110,3 @@ private:
 	sigc::connection m_connRotationDampingToggleKey;
 	sigc::connection m_fireMissileKey;
 };
-
-#endif
