+++ src/ship/ShipViewController.cpp
@@ -1,3 +1,5 @@
+// Copyright © 2008-2019 Pioneer Developers. See AUTHORS.txt for details
+// Licensed under the terms of the GPL v3. See licenses/GPL-3.txt
 
 #include "ShipViewController.h"
 #include "CameraController.h"
